Add books pipeline

+++ Includes/Copy-Datasets.py
@@ -1,13 +1,25 @@
 # Databricks notebook source
+def path_exists(path):
+  try:
+    dbutils.fs.ls(path)
+    return True
+  except Exception as e:
+    if 'java.io.FileNotFoundException' in str(e):
+      return False
+    else:
+      raise
+
+# COMMAND ----------
+
 def download_dataset(source, target):
     files = dbutils.fs.ls(source)
 
     for f in files:
-        print(f"Copying {f.name} ...")
         source_path = f"{source}/{f.name}"
         target_path = f"{target}/{f.name}"
-
-        dbutils.fs.cp(source, target, True)
+        if not path_exists(target_path):
+            print(f"Copying {f.name} ...")
+            dbutils.fs.cp(source, target, True)
 
 # COMMAND ----------
 
@@ -17,9 +29,6 @@ spark.conf.set(f"dataset.bookstore", dataset_bookstore)
 
 # COMMAND ----------
 
-streaming_dir = f"{dataset_bookstore}/orders-streaming"
-raw_dir = f"{dataset_bookstore}/orders-raw"
-
 def get_index(dir):
     files = dbutils.fs.ls(dir)
     index = 0
@@ -27,7 +36,12 @@ def get_index(dir):
         file = max(files).name
         index = int(file.rsplit('.', maxsplit=1)[0])
     return index+1
-        
+
+# COMMAND ----------
+
+# Structured Streaming
+streaming_dir = f"{dataset_bookstore}/orders-streaming"
+raw_dir = f"{dataset_bookstore}/orders-raw"
 
 def load_file(current_index):
     latest_file = f"{str(current_index).zfill(2)}.parquet"
@@ -50,17 +64,34 @@ def load_new_data(all=False):
 
 # COMMAND ----------
 
-def path_exists(path):
-  try:
-    dbutils.fs.ls(path)
-    return True
-  except Exception as e:
-    if 'java.io.FileNotFoundException' in str(e):
-      return False
+# DLT
+streaming_orders_dir = f"{dataset_bookstore}/orders-json-streaming"
+streaming_books_dir = f"{dataset_bookstore}/books-streaming"
+
+raw_orders_dir = f"{dataset_bookstore}/orders-json-raw"
+raw_books_dir = f"{dataset_bookstore}/books-cdc"
+
+def load_json_file(current_index):
+    latest_file = f"{str(current_index).zfill(2)}.json"
+    print(f"Loading {latest_file} orders file to the bookstore dataset")
+    dbutils.fs.cp(f"{streaming_orders_dir}/{latest_file}", f"{raw_orders_dir}/{latest_file}")
+    print(f"Loading {latest_file} books file to the bookstore dataset")
+    dbutils.fs.cp(f"{streaming_books_dir}/{latest_file}", f"{raw_books_dir}/{latest_file}")
+
+    
+def load_new_json_data(all=False):
+    index = get_index(raw_orders_dir)
+    if index >= 10:
+        print("No more data to load\n")
+
+    elif all == True:
+        while index <= 10:
+            load_json_file(index)
+            index += 1
     else:
-      raise
+        load_json_file(index)
+        index += 1
 
 # COMMAND ----------
 
-if not path_exists(dataset_bookstore):
-    download_dataset(data_source_uri, dataset_bookstore)
+download_dataset(data_source_uri, dataset_bookstore)
